# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JISHLink workforce management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List pending submissions
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PendingSubmission"
  /submissions/{id}/approve:
    post:
      operationId: approveSubmission
      tags: [submissions]
      summary: Approve a submission and create employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveSubmissionRequest"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /submissions/{id}/reject:
    post:
      operationId: rejectSubmission
      tags: [submissions]
      summary: Reject a submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectSubmissionRequest"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Employee:
      type: object
      properties:
        id:
          type: string
        employee_code:
          type: string
        full_name:
          type: string
        dob:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        aadhar_number:
          type: string
          nullable: true
        pan_number:
          type: string
          nullable: true
        qualification:
          type: string
          nullable: true
        marital_status:
          type: string
          nullable: true
        blood_group:
          type: string
          nullable: true
        contact_number:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        emergency_contact:
          type: string
          nullable: true
        nominee_name:
          type: string
          nullable: true
        nominee_relation:
          type: string
          nullable: true
        designation:
          type: string
          nullable: true
        workplace_id:
          type: string
          nullable: true
        workplace:
          $ref: "#/components/schemas/Workplace"
          nullable: true
        date_of_joining:
          type: string
          nullable: true
        date_of_leaving:
          type: string
          nullable: true
        employment_status:
          type: string
          nullable: true
        employment_type:
          type: string
          nullable: true
        reporting_manager_id:
          type: string
          nullable: true
        pf_number:
          type: string
          nullable: true
        esi_number:
          type: string
          nullable: true
        uan_number:
          type: string
          nullable: true
        bank_name:
          type: string
          nullable: true
        bank_branch:
          type: string
          nullable: true
        account_number:
          type: string
          nullable: true
        ifsc_code:
          type: string
          nullable: true
        driving_license_number:
          type: string
          nullable: true
        vehicle_details:
          type: string
          nullable: true
        photo_url:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
        password_changed:
          type: boolean
          nullable: true
        created_at:
          type: string
          nullable: true
      required: [id, employee_code, full_name]
...
    Workplace:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        client_name:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
      required: [id, name]
...
    PendingSubmission:
      type: object
      properties:
        id:
          type: string
        source:
          type: string
        submitted_data:
          type: object
          additionalProperties: true
        validation_results:
          type: object
          additionalProperties: true
          nullable: true
        status:
          type: string
        admin_remarks:
          type: string
          nullable: true
        submitted_at:
          type: string
      required: [id, source, submitted_data, status, submitted_at]
    ApproveSubmissionRequest:
      type: object
      properties:
        workplace_id:
          type: string
        reporting_manager_id:
          type: string
        username:
          type: string
        password:
          type: string
        designation:
          type: string
          nullable: true
      required: [workplace_id, reporting_manager_id, username, password]
    RejectSubmissionRequest:
      type: object
      properties:
        remarks:
          type: string
      required: [remarks]
...