# lib/api-zod/src/generated/types/pendingSubmission.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JISHLink workforce management API
 * OpenAPI spec version: 0.1.0
 */
import type { PendingSubmissionSubmittedData } from './pendingSubmissionSubmittedData';
import type { PendingSubmissionValidationResults } from './pendingSubmissionValidationResults';

export interface PendingSubmission {
  id: string;
  source: string;
  submitted_data: PendingSubmissionSubmittedData;
  validation_results?: PendingSubmissionValidationResults;
  status: string;
  admin_remarks?: string | null;
  submitted_at: string;
}